changed accrual type from int to float64

+++ internal/app/entity/order.go
@@ -7,6 +7,6 @@ type OrderNumber string
 type Order struct {
 	Number      OrderNumber
 	Status      string
-	Accrual     int
+	Accrual     float64
 	DateCreated string
 }
