+++ internal/app/model/accrual.go
@@ -1,7 +1,7 @@
 package model
 
 type AccrualResponse struct {
-	Number  string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual,omitempty"`
+	Number  string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual,omitempty"`
 }
