+++ internal/app/model/order.go
@@ -21,8 +21,8 @@ const (
 type UploadedOrders []UploadedOrder
 
 type UploadedOrder struct {
-	Number     string `json:"number"`
-	Status     string `json:"status"`
-	Accrual    int    `json:"accrual"`
-	UploadTime string `json:"uploaded_at"`
+	Number     string  `json:"number"`
+	Status     string  `json:"status"`
+	Accrual    float64 `json:"accrual"`
+	UploadTime string  `json:"uploaded_at"`
 }
